upgrade

+++ node_middleware/Wrapper/methods/helloLoop.js
@@ -7,22 +7,35 @@ const proto = props.proto;
 const sessionVar = props.sessionVar;
 
 function helloLoop() {
-  // Added by Alex, for the purpose of Pipeline Visulization
   console.log("helloLoop begin");
-  let pathPrefix = "responses/";
-  if (fs.existsSync(pathPrefix)) {
-    console.log("Remove old responses folder!");
-    fse.removeSync(pathPrefix);
+
+  if (props.isRequest) {
+    let pathPrefix = props.REQUESTS_PATH;
+    if (fs.existsSync(pathPrefix)) {
+      console.log("Remove old request folder!");
+      fse.removeSync(pathPrefix);
+    }
+    console.log("Create new request folder!!");
+    fs.mkdirSync(pathPrefix);
   }
-  console.log("Create a new responses folder!!");
-  fs.mkdirSync(pathPrefix);
+
+  if (props.isResponse) {
+    let pathPrefix = props.RESPONSES_PATH;
+    if (fs.existsSync(pathPrefix)) {
+      console.log("Remove old responses folder!");
+      fse.removeSync(pathPrefix);
+    }
+    console.log("Create new responses folder!!");
+    fs.mkdirSync(pathPrefix);
+  }
 
   let promise = new Promise((fulfill, reject) => {
     let request = new proto.HelloRequest();
     let waiting = false;
     setInterval(() => {
-      // let sessionVar = props.sessionVar;
-      if (waiting || sessionVar.connected) return;
+      if (waiting || sessionVar.connected) {
+        return;
+      }
       waiting = true;
       let client = props.client;
       client.Hello(request, (err, response) => {
@@ -33,18 +46,21 @@ function helloLoop() {
           // we do not reject here, because ta2 can becaome available at some point
           // reject(err);
         } else {
+          console.log("Success!Hello");          
           sessionVar.connected = true;
-          console.log("Success!Hello", response);
           sessionVar.ta2Ident = response;
+          // props.allowed_val_types = response.allowed_value_types; 
           fulfill(sessionVar);
 
           // Added by Alex, for the purpose of Pipeline Visulization
-          let pathPrefix = "responses/";
-          let pathMid = "helloResponse";
-          let pathAffix = ".json";
-          let path = pathPrefix + pathMid + pathAffix;
-          let responseStr = JSON.stringify(response);
-          fs.writeFileSync(path, responseStr);
+          if (props.isResponse) {
+            let pathPrefix = props.RESPONSES_PATH;
+            let pathMid = "helloResponse";
+            let pathAffix = ".json";
+            let path = pathPrefix + pathMid + pathAffix;
+            let responseStr = JSON.stringify(response);
+            fs.writeFileSync(path, responseStr);
+          }
         }
       });
     }, 10000);
